IP_STACK: v6 - allow MIRROR_IMAGES=false
IPv6 forces MIRROR_IMAGES and sets disableAllDefaultSources for
OperatorHub, and the pull-secret contains only an entry for the
mirror. This probably makes sense for disconnected installs,
but not when there is actual connectivity.

This changes the code to only set MIRROR_IMAGES=true when it is
undefined, and all conditions on MIRROR_IMAGES has been updated to
treat the value `false` the same way it treats unset.

This should allow to explicitly disable MIRROR_IMAGES, and thus allow
deploying with IP_STACK=v6 without the image mirror.